+++ src/pages/chat/sections/chatMessageList/ChatMessageList.tsx
@@ -1,6 +1,8 @@
 import { useEffect, useMemo, useState } from 'react';
 import { Virtuoso } from 'react-virtuoso';
 
+import { toDateKey, toFiveMinutesKey, toKstDate } from '@/utils/chat.utils';
+
 import type {
   ChatTypes,
   GroupedChatListTypes,
@@ -25,24 +27,6 @@ const renderDataByTime = (
 
 function ChatMessageList() {
   const [atBottom, setAtBottom] = useState(true);
-  const toKstDate = (iso: string | Date) => {
-    const kst = typeof iso === 'string' ? Date.parse(iso) : iso.getTime();
-    return new Date(kst + 9 * 60 * 60 * 1000);
-  };
-
-  const toDateKey = (dt: Date) => {
-    const year = dt.getUTCFullYear();
-    const month = String(dt.getUTCMonth() + 1).padStart(2, '0');
-    const day = String(dt.getUTCDate()).padStart(2, '0');
-    return `${year}-${month}-${day}`;
-  };
-
-  const toFiveMinutesKey = (dt: Date) => {
-    const hour = String(dt.getUTCHours()).padStart(2, '0');
-    const minute = dt.getUTCMinutes();
-    const flooredMinute = String(Math.floor(minute / 5) * 5).padStart(2, '0');
-    return `${hour}:${flooredMinute}`;
-  };
 
   const sortedChatData = [...CHAT_EXAMPLES].sort(
     (a, b) => new Date(a.sendAt).getTime() - new Date(b.sendAt).getTime(),
